Changes and bug fixes
- changed color of asteroids
- added sprites to caves on main asteroid scene
- made changes to some scripts where calls are made dynamically and not dependable on serializable fields
- fixed bug where camera not following the player when entering the cave; player was not spawning as default spawn.
- fixed bug where player manager was not remembering the last coordinates of the character
