+++ Assets/Scripts/Player/PlayerManager.cs
@@ -25,6 +25,7 @@ public class PlayerManager : MonoBehaviour
 
     void OnSceneLoaded(Scene scene, LoadSceneMode mode) 
     {
+        Debug.Log(this.lastCoordinates);
         playerInstance = GameObject.FindWithTag("Player");
         playerController = playerInstance.GetComponent<PlayerController>();  
         setScenePosition(); 
@@ -38,20 +39,6 @@ public class PlayerManager : MonoBehaviour
         cam.transform.rotation = Quaternion.Euler(0,0,0);
     }
 
-    // Start is called before the first frame update
-    void Awake()
-    {
-        // // Singleton method
-        if (instance != null && instance != this)
-        {
-            Destroy(gameObject);
-        }
-        else 
-        {
-            instance = this; 
-            DontDestroyOnLoad(gameObject);
-        }
-    }
 
     // set the players initial position and jump force
     private void setScenePosition() 
@@ -68,7 +55,7 @@ public class PlayerManager : MonoBehaviour
                     this.playerInstance.transform.position = this.lastCoordinates; 
                     this.playerInstance.transform.rotation = this.lastRotation;
                     this.playerController.UpdateJumpForce(1.5f);
-                    Debug.Log(this.playerController.GetJumpForce());
+                    Debug.Log("1");
                 }
                 else 
                 {
@@ -77,7 +64,7 @@ public class PlayerManager : MonoBehaviour
                     this.playerInstance.transform.position = defaultSpawn.position;
                     this.playerInstance.transform.rotation = Quaternion.Euler(0,0,0);
                     this.playerController.UpdateJumpForce(1.5f);
-                    Debug.Log(this.playerController.GetJumpForce());
+                    Debug.Log("2");
                 }
                 break;
 
@@ -87,7 +74,7 @@ public class PlayerManager : MonoBehaviour
                 this.playerInstance.transform.position = defaultSpawn.position;
                 this.playerInstance.transform.rotation = Quaternion.Euler(0,0,0);
                 this.playerController.UpdateJumpForce(0.5f);
-                Debug.Log(this.playerController.GetJumpForce());
+                Debug.Log("3");
                 break; 
         }
     }
