+++ ProjectSettings/EditorBuildSettings.asset
@@ -15,7 +15,7 @@ EditorBuildSettings:
     path: Assets/Scenes/AsteroidScene.unity
     guid: ebf1ab7c3b679a54396d68b3f4eaea7d
   - enabled: 1
-    path: Assets/Scenes/CaveScenes/Cave_1.unity
+    path: Assets/Scenes/CaveScenes/Cave 1.unity
     guid: 40c97177fd07f9644adcbf669ac7a274
   - enabled: 1
     path: Assets/Scenes/CaveScenes/Cave_2.unity
