+++ ProjectSettings/TagManager.asset
@@ -7,8 +7,9 @@ TagManager:
   - Background
   - VirtualPlayerCamera
   - PlayerLastCoordinates
-  - PlayerManager
   - Ladder
+  - GameManager
+  - Caves
   layers:
   - Default
   - TransparentFX
